Split up ImageZoom into multiple compoents
