--- packages/components/src/image_zoom/constants.jsx
@@ -1,4 +1,0 @@
-export const MAX_SCALE_FACTOR = 4;
-export const CONTAINER_WIDTH_CHANGE_RATE = 500;
-export const DOUBLE_TAP_TIMEOUT = 300;
-export const DOUBLE_TAP_THRESHOLD = 20;
