+++ packages/components/src/image_zoom_provider/constants.jsx
@@ -1,0 +1,1 @@
+export const MAX_SCALE_FACTOR = 4;
