+++ packages/components/src/image_zoom_provider/context.jsx
@@ -1,0 +1,8 @@
+import { createContext } from 'react';
+
+const Context = createContext();
+
+const { Provider } = Context;
+
+export { Provider };
+export default Context;
