+++ packages/components/src/image_zoom_slider/hooks.jsx
@@ -1,0 +1,9 @@
+import { useEffect } from 'react';
+
+export const useStateControlledInput = (ref, state) => {
+  useEffect(() => {
+    if (!ref.current) return;
+
+    ref.current.setValue(state, null, { silent: true });
+  }, [state]);
+};
