+++ packages/components/src/image_zoom_view/constants.jsx
@@ -1,0 +1,3 @@
+export const DOUBLE_TAP_TIMEOUT = 300;
+export const DOUBLE_TAP_THRESHOLD = 20;
+export const CONTAINER_WIDTH_CHANGE_RATE = 500;
